fixed plt float bug

+++ Data_extract/utils.py
@@ -868,8 +868,8 @@ def align_data(
                 column_names[0]
             ].unique()[0]
 
-            t = data_data[:, 0]
-            y = data_data[:, 1]
+            t = data_data[:, 0].astype('float64')
+            y = data_data[:, 1].astype('float64')
             t_known = t[~np.isnan(y)]
             y_known = y[~np.isnan(y)]
             t_missing = t[np.isnan(y)]
